moved function runner to fnruntime
